+++ lib/src/cli_opt.dart
@@ -69,6 +69,8 @@ class CliOpt {
     }
   }
 
+  /// Standard serializer
+  ///
   @override
   String toString() {
     return '"$fullName": $values';