+++ lib/src/list_ext.dart
@@ -4,5 +4,7 @@
 /// Extension methods to manipulate generic lists
 ///
 extension ListEx on List {
+  /// A method to get thye first element or null (if the list is empty)
+  ///
   dynamic firstOrDefault({dynamic defValue}) => (isEmpty ? defValue : first);
 }